pyproject.toml
[project]
name = "3dprinter-error-detection"
version = "0.1.0"
description = "Code for Cameron Palmer's 3D printing error detection experiment, April 2023."
authors = [
    {name = "Cameron Palmer",email = "cameron@cameronmpalmer.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "tqdm (>=4.67.1,<5.0.0)",
    "wandb (>=0.19.9,<0.20.0)",
    "octorest (>=0.4,<0.5)",
    "torch (>=2.6.0,<3.0.0)",
    "torchvision (>=0.21.0,<0.22.0)",
    "black (>=25.1.0,<26.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "pytorch-lightning (>=2.5.1,<3.0.0)",
    "matplotlib (>=3.10.1,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
